moved rave actions into home controller
