+++ app/controllers/rave_controller.rb
@@ -1,0 +1,12 @@
+class RaveController < ActionController::Base
+  layout 'application'
+
+  def index
+    @status = Rails.cache.read("rave_status")
+
+    respond_to do |format|
+      format.rss { render :layout => false }
+    end
+  end
+
+end
