--- app/views/rave_simulation/index.html.erb
@@ -1,52 +1,0 @@
-<div class="rave-simulator">
-  <div class="rave-status">
-  
-  <% if @status == "on" %>
-    <p class="rave-on">ON</p>
-  <% else %>
-    <p class="rave-off">OFF</p>
-  <% end %>
-  </div>
-  <h3>Rave Simulator</h3>
-  <div class="toggle-container">
-    <% if @status == "on" %>
-      <% button_text = "Turn Off" %>
-    <% else %>
-      <% button_text = "Turn On" %>
-    <% end %>
-
-    <%= link_to button_text, {:controller=> "rave_simulation", :action => "send_alert", :status => @status}, method: :post, class: 'btn', local: true %>
-    
-  </div>
-
-  <div class="rave-feed">
-    <h3>Simulated Rave Feed</h3>
-    <table class="feed-table">
-      <% @simulated_rss.items.each do |item| %>
-        <tr>
-          <th class="time-field"><%= item.updated.content.to_datetime.strftime('%F at %l:%M %P') %></th>
-          <th>
-            <p><%= item.title.content %></p>
-            <p><%= item.summary.content %></p>
-          </th>
-        </tr>
-      <% end %>
-    </table>
-  </div>
-
-  <div class="rave-feed">
-    <h3>Live Rave Feed</h3>
-    <table class="feed-table">
-      <% @live_rss.items.each do |item| %>
-        <tr>
-          <th class="time-field"><%= item.pubDate.to_datetime.strftime('%F at %l:%M %P') %></th>
-          <th>
-            <p><%= item.title %></p>
-            <p><%= item.description %></p>
-          </th>
-        </tr>
-      <% end %>
-    </table>
-  </div>
-  <%= link_to "Back to Home Page", root_path, class: "nav-link" %>
-</div>
